+++ Backend/middleware/imageDelete.js
@@ -1,0 +1,6 @@
+const fs = require('fs');
+const { promisify } = require('util');
+
+const unlinkAsync = promisify(fs.unlink);
+
+module.exports = unlinkAsync;
